feat: revamp handling JWT

+++ R.Systems.Auth.Core/Interfaces/IUserWriteRepository.cs
@@ -5,6 +5,6 @@ namespace R.Systems.Auth.Core.Interfaces
 {
     public interface IUserWriteRepository
     {
-        Task SaveRefreshTokenAsync(long userId, string refreshToken, int lifetimeInMinutes);
+        Task SaveRefreshTokenAsync(long userId, string refreshToken, double lifetimeInMinutes);
     }
 }
